apps/website: first steps

+++ packages/pema/src/private/BigIntDataType.ts
@@ -1,3 +1,3 @@
-type BigintDataType = "i64" | "i128";
+type BigintDataType = "i128" | "i64";
 
 export type { BigintDataType as default };
