+++ packages/pema/src/private/UintDataType.ts
@@ -1,3 +1,3 @@
-type UintDataType = "u8" | "u16" | "u32";
+type UintDataType = "u16" | "u32" | "u8";
 
 export type { UintDataType as default };
